requirements.txt
streamlit
pypdf
pillow